[XrdHttpTpc] Rename everything from XrdTpc to XrdHttpTpc
This is to make the headers, source files, namespaces, etc all the
same name as the plugin itself, for consistency.
